+++ engineer/plugins/core.py
@@ -32,7 +32,7 @@ def load_plugins():
 
 
 def get_all_plugin_types():
-    return ThemeProvider, PostProcessor, CommandPlugin, FilterPlugin
+    return ThemeProvider, PostProcessor, CommandPlugin, JinjaEnvironmentPlugin
 
 
 #noinspection PyMissingConstructor,PyUnusedLocal
@@ -224,26 +224,90 @@ class CommandPlugin(PluginMixin):
         raise NotImplementedError()
 
 
-class FilterPlugin(PluginMixin):
+class JinjaEnvironmentPlugin(PluginMixin):
     """
-    Base class for filter plugins.
+    Base class for JinjaEnvironment :ref:`plugins`.
+
+    JinjaEnvironment plugins can supplement the Jinja 2 environment with things like filters and global
+    functions. These additions can then be used in your Jinja templates.
+
+    .. versionadded:: 0.5.0
     """
     __metaclass__ = PluginMount
-    
-    filters = ()
-    
+
+    filters = {}
+    """
+    A dict of filters to add to the Jinja environment. The key of each entry should be the name of the filter (as it
+    will be used inside templates), while the value should be the filter function. If you require more custom logic
+    to build the dict of filters, override the :meth:`~engineer.plugins.JinjaEnvironmentPlugin.get_filters` method.
+    """
+
+    globals = {}
+    """
+    A dict of functions to add to the Jinja environment globally. The key of each entry should be the name of the
+    function (as it will be used inside templates), while the value should be the function itself. If you require more
+    custom logic to build this dict, override the :meth:`~engineer.plugins.JinjaEnvironmentPlugin.get_globals` method.
+    """
+
     @classmethod
-    def add_filters(cls, jinja_env):
+    def _add_filters(cls, jinja_env):
         logger = cls.get_logger()
-        filter_list = cls.get_filters()
-        for filter in filter_list:
-            filter_name = filter.__name__
+        filters = cls.get_filters()
+        for filter_name, filter_function in filters.iteritems():
             if filter_name in jinja_env.filters:
-                logger.warning("Filter name conflict. A plugin is trying to add a filter with a name that conflicts with an existing filter. Filter name: %s" % filter_name)
+                logger.warning("Jinja filter name conflict. "
+                               "A plugin is trying to add a filter with a name that conflicts with an existing filter. "
+                               "Filter name: %s" % filter_name)
             else:
-                jinja_env.filters[filter_name] = filter
-                logger.debug("Registered filter: %s" % filter_name)
-    
+                jinja_env.filters[filter_name] = filter_function
+                logger.debug("Registered Jinja filter: %s" % filter_name)
+
+    @classmethod
+    def _add_globals(cls, jinja_env):
+        logger = cls.get_logger()
+        global_list = cls.get_globals()
+        for global_name, the_global in global_list.iteritems():
+            if global_name in jinja_env.globals:
+                logger.warning("Jinja global name conflict. "
+                               "A plugin is trying to add a global with a name that conflicts with an existing global. "
+                               "Global name: %s" % global_name)
+            else:
+                jinja_env.globals[global_name] = the_global
+                logger.debug("Registered Jinja global: %s" % global_name)
+
+    @classmethod
+    def update_environment(cls, jinja_env):
+        """
+        For complete customization of the Jinja environment, subclasses can override this method.
+
+        Subclasses should ensure that the base implementation is called first in their overridden implementation. For
+        example:
+
+        .. code-block:: python
+
+            @classmethod
+            def update_environment(cls, jinja_env):
+                super(BundledFilters, cls).update_environment(jinja_env)
+                # some other code here...
+
+        :param jinja_env: The Jinja environment.
+        """
+        cls._add_filters(jinja_env)
+        cls._add_globals(jinja_env)
+
     @classmethod
     def get_filters(cls):
-        return filters
+        """
+        If required, subclasses can override this method to return a dict of filters to add to the Jinja environment.
+        The default implementation simply returns :attr:`~engineer.plugins.JinjaEnvironmentPlugin.filters`.
+        """
+        return cls.filters
+
+    @classmethod
+    def get_globals(cls):
+        """
+        If required, subclasses can override this method to return a dict of functions to add to the Jinja
+        environment globally. The default implementation simply
+        returns :attr:`~engineer.plugins.JinjaEnvironmentPlugin.globals`.
+        """
+        return cls.globals