+++ engineer/docs_source/changelog.rst
@@ -19,6 +19,10 @@ version 0.5.0 -
 - New plugin added to support :ref:`Markdown Lazy Links<lazy links plugin>`. This plugin is enabled by default,
   so you can start using lazy links immediately.
 - New 'experimental' plugin permissions model. More developer information at: :ref:`plugin permissions`.
+- Jinja environment can now be modified using :class:`~engineer.plugins.JinjaEnvironmentPlugin`.
+- Jinja syntax can now be used in post content thanks to the new :ref:`jinja post processor plugin`.
+- A new tag for :ref:`inserting images into posts<post images>`. The output can
+  be :ref:`customized by themes<image template>`.
   
 
 version 0.4.5 - October 2, 2013